# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Channel:
    properties:
      avatar:
        type: string
      birthday:
        type: string
      channelUrl:
        type: string
      company:
        type: string
      createdAt:
        type: string
      debutDate:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      height:
        type: number
      id:
        type: integer
      name:
        type: string
      twitterAccount:
        type: string
      unit:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: Oshi-api swagger
  license:
    name: MIT
    url: https://www.mit.edu/~amini/LICENSE.md
  title: Oshi-api
  version: "1.0"
paths:
  /api/v1/channels:
    get:
      consumes:
      - application/json
      description: Get channels by filters
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Channel'
            type: array
      summary: Get Channels
      tags:
      - Channels
    post:
      consumes:
      - application/json
      description: Store new channel
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
      summary: Create Channl
      tags:
      - Channels
  /api/v1/channels/{twitterAccount}:
    get:
      consumes:
      - application/json
      description: Show channel info
      parameters:
      - description: Twitter Account
        in: path
        name: twitterAccount
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
        "404":
          description: Not Found
          schema:
            type: string
      tags:
      - Channels
swagger: "2.0"